requirements.txt
absl-py==0.11.0
appdirs==1.4.3
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
build==0.4.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
check-manifest==0.46
colorama==0.4.3
colorlog==4.4.0
contextlib2==0.6.0
cycler==0.10.0
Cython==0.29.23
dataclasses==0.6
decorator==4.4.2
defusedxml==0.7.1
DicomRTTool==0.3.13
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imageio==2.9.0
importlib-metadata==2.0.0
iniconfig==1.1.1
ipaddr==2.2.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mistune==0.8.4
mpmath==1.2.1
msgpack==0.6.2
mss==6.0.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
nilabels==0.0.8
nilearn==0.8.0
notebook==6.4.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.3
pafy==0.5.5
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.13.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydicom==2.1.2
Pygments==2.9.0
pyparsing==2.4.6
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
rt-utils==1.1.4
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.1
Send2Trash==1.5.0
SimpleITK==2.0.2
six==1.14.0
sklearn==0.0
sympy==1.8
tabulate==0.8.9
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.6.14
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
vidgear==0.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
yacs==0.1.8
youtube-dl==2020.9.20
zipp==3.4.0
